requirements.txt
asgiref
channels
channels_redis
daphne
django
django-allauth
django-auth-ldap
django-eventstream
django-webpack-loader
maxminddb
netCDF4
numpy
oauthlib
pyparsing
psycopg2-binary
radar-data
setproctitle
tqdm
Twisted[tls,http2]
websocket_client
